Modify ObjectGroup underlying type
To reduce the amount of data to be stored in map files.

Also, add NONE entry to indicate empty group.

+++ src/fheroes2/editor/editor_interface.h
@@ -36,7 +36,7 @@ namespace Maps
 {
     class Tiles;
 
-    enum class ObjectGroup : int32_t;
+    enum class ObjectGroup : uint8_t;
 }
 
 namespace Interface
